% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excitr_keys_to_xref.R
\name{excitr_assemble}
\alias{excitr_assemble}
\title{Assemble Multiple Citations with APA Formatting}
\usage{
excitr_assemble(citations)
}
\arguments{
\item{citations}{A vector of citation strings.}
}
\value{
A single string with all citations separated by semicolons.
}
\description{
Assembles a list of citations into a single inline reference, using semicolons.
}
\seealso{
Other cite: 
\code{\link{excitr_bib_keys_missing}()},
\code{\link{excitr_format}()},
\code{\link{excitr_format_single}()},
\code{\link{excitr_keys_extract}()},
\code{\link{excitr_keys_to_inline}()},
\code{\link{excitr_keys_to_xref}()}
}
\concept{cite}
