% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excitr_bib_clean.R
\name{excitr_bib_keys_missing}
\alias{excitr_bib_keys_missing}
\alias{excitr_bib_keys_extract}
\alias{excitr_keys_guess_match}
\title{Check for Missing Citation Keys in a BibTeX File}
\usage{
excitr_bib_keys_missing(path_bib, citations)

excitr_bib_keys_extract(path_bib)

excitr_keys_guess_match(keys_missing, keys_bib)
}
\arguments{
\item{path_bib}{A file path to the bibliography file (\code{.bib}) used for citation keys.
Defaults to \code{"references.bib"} in the current working directory.}

\item{keys}{\link{character} A vector of citation keys to check for in the BibTeX file.}
}
\value{
A vector of citation keys missing from the BibTeX file. Returns an empty
vector if all citation keys are found.

A character vector of citation keys extracted from the BibTeX file.

A character vector of citation keys extracted from the BibTeX file.
}
\description{
This function identifies citation keys provided in a list that are missing from
a specified BibTeX file. It ensures the BibTeX file exists and processes the
file to match the citation keys.

This function extracts all citation keys from a given BibTeX file.
}
\examples{
\dontrun{
path_bib <- system.file("extdata", "references.bib", package = "ngr")
citations <- c("Smith2020", "Jones2019")
excitr_bib_keys_missing(path_bib, citations)
}
}
\seealso{
Other cite: 
\code{\link{excitr_assemble}()},
\code{\link{excitr_format}()},
\code{\link{excitr_format_single}()},
\code{\link{excitr_keys_extract}()},
\code{\link{excitr_keys_to_inline}()},
\code{\link{excitr_keys_to_xref}()}
}
\concept{cite}
