% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excitr_keys_to_xref.R
\name{excitr_format}
\alias{excitr_format}
\title{Format Multiple Citations}
\usage{
excitr_format(raw_key, bib_obj, ...)
}
\arguments{
\item{raw_key}{A raw citation key string, possibly with brackets.}

\item{bib_obj}{The bibliography object.}

\item{...}{Additional arguments passed from the main function.}
}
\value{
Formatted citation as a string.
}
\description{
Formats one or more citation keys into a single inline reference.
}
\seealso{
Other cite: 
\code{\link{excitr_assemble}()},
\code{\link{excitr_bib_keys_missing}()},
\code{\link{excitr_format_single}()},
\code{\link{excitr_keys_extract}()},
\code{\link{excitr_keys_to_inline}()},
\code{\link{excitr_keys_to_xref}()}
}
\concept{cite}
