% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excitr_keys_to_xref.R
\name{excitr_format_single}
\alias{excitr_format_single}
\title{Format a Single Citation for Inline Text}
\usage{
excitr_format_single(key, bib_obj, key_abort_on_missing = TRUE, ...)
}
\arguments{
\item{key}{The citation key.}

\item{bib_obj}{The bibliography object created by reading in a .bib file with \code{RefManageR::ReadBib}.}

\item{key_abort_on_missing}{Logical. If TRUE, aborts on missing citation keys.}

\item{...}{Additional arguments passed from the main function.}
}
\value{
Formatted citation as a string.
}
\description{
Formats a single citation key into inline text without brackets.
}
\seealso{
Other cite: 
\code{\link{excitr_assemble}()},
\code{\link{excitr_bib_keys_missing}()},
\code{\link{excitr_format}()},
\code{\link{excitr_keys_extract}()},
\code{\link{excitr_keys_to_inline}()},
\code{\link{excitr_keys_to_xref}()}
}
\concept{cite}
