% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_cite_keys_to_xref.R
\name{ngr_cite_assemble}
\alias{ngr_cite_assemble}
\title{Assemble Multiple Citations with APA Formatting}
\usage{
ngr_cite_assemble(citations)
}
\arguments{
\item{citations}{A vector of citation strings.}
}
\value{
A single string with all citations separated by semicolons.
}
\description{
Assembles a list of citations into a single inline reference, using semicolons.
}
\seealso{
Other cite: 
\code{\link{ngr_cite_format}()},
\code{\link{ngr_cite_format_single}()},
\code{\link{ngr_cite_keys_to_xref}()}
}
\concept{cite}
