% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_cite_bib_clean.R
\name{ngr_cite_bib_keys_missing}
\alias{ngr_cite_bib_keys_missing}
\alias{ngr_cite_bib_keys_extract}
\alias{ngr_cite_keys_guess_match}
\title{Check for Missing Citation Keys in a BibTeX File}
\usage{
ngr_cite_bib_keys_missing(path_bib, citations)

ngr_cite_bib_keys_extract(path_bib)

ngr_cite_keys_guess_match(keys_missing, keys_bib)
}
\arguments{
\item{path_bib}{A string specifying the path to the bibliography file (e.g., \code{references.bib}).}

\item{keys}{\link{character} A vector of citation keys to check for in the BibTeX file.}
}
\value{
A vector of citation keys missing from the BibTeX file. Returns an empty
vector if all citation keys are found.

A character vector of citation keys extracted from the BibTeX file.

A character vector of citation keys extracted from the BibTeX file.
}
\description{
This function identifies citation keys provided in a list that are missing from
a specified BibTeX file. It ensures the BibTeX file exists and processes the
file to match the citation keys.

This function extracts all citation keys from a given BibTeX file.
}
\examples{
\dontrun{
path_bib <- system.file("extdata", "references.bib", package = "ngr")
citations <- c("Smith2020", "Jones2019")
ngr_cite_bib_keys_missing(path_bib, citations)
}
}
\seealso{
Other cite: 
\code{\link{ngr_cite_assemble}()},
\code{\link{ngr_cite_format}()},
\code{\link{ngr_cite_format_single}()},
\code{\link{ngr_cite_keys_extract}()},
\code{\link{ngr_cite_keys_to_inline_table_col}()},
\code{\link{ngr_cite_keys_to_xref}()}
}
\concept{cite}
