% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_cite_keys_extract.R
\name{ngr_cite_keys_extract}
\alias{ngr_cite_keys_extract}
\alias{ngr_cite_keys_extract_table}
\title{Extract Citations from Text}
\usage{
ngr_cite_keys_extract(text, keys_additional = NULL)

ngr_cite_keys_extract_table(dat, col_format, print_rowwise = TRUE, ...)
}
\arguments{
\item{text}{A character string containing the text to process. Passing \code{NA} will issue a warning.}

\item{keys_additional}{A character vector of additional citation keys to include. Default is \code{NULL}.}

\item{dat}{A data frame containing the column to transform.}

\item{col_format}{A string specifying the name of the column to transform.}

\item{print_rowwise}{Logical. If \code{TRUE}, formats and prints keys row-wise with a trailing
comma for easy YAML compatibility. Defaults to \code{TRUE}.}

\item{...}{Can be used to pass additional citation keys to  \code{\link[=ngr_cite_keys_extract]{ngr_cite_keys_extract()}} in case there are
citation keys additional to the ones in the table that you want in the output.}
}
\value{
A character vector of unique citation keys.

A character vector of citation keys extracted from a data frame column.
}
\description{
This function extracts citations in the form of \verb{@citation} or \verb{[e.g., @citation1; @citation2]}
from a given text. Additional citations can also be included via a parameter.

This function extracts citation keys from a specified column in a data frame
and optinally formats them as a character vector or prints them row-wise for easy copy-pasting
into the \code{nocite} line of YAML in an R Markdown document.  This can be handy when we are using
\code{\link[=ngr_cite_keys_extract]{ngr_cite_keys_extract()}} to format a table column for reporting but we
require the citation keys so that they show up in the references section of the overall report.
}
\details{
This function builds on \code{\link[=ngr_cite_keys_extract]{ngr_cite_keys_extract()}}, and integrates its functionality
for working with citation keys extracted from a data frame.
}
\seealso{
\code{\link[=ngr_cite_keys_extract]{ngr_cite_keys_extract()}}

Other cite: 
\code{\link{ngr_cite_assemble}()},
\code{\link{ngr_cite_bib_keys_missing}()},
\code{\link{ngr_cite_format}()},
\code{\link{ngr_cite_format_single}()},
\code{\link{ngr_cite_keys_to_inline}()},
\code{\link{ngr_cite_keys_to_xref}()}
}
\concept{cite}
