% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_cite_keys_to_inline_table_col.R
\name{ngr_cite_keys_to_inline_table_col}
\alias{ngr_cite_keys_to_inline_table_col}
\title{Transform Citation Keys in a Data Frame Column}
\usage{
ngr_cite_keys_to_inline_table_col(
  dat = NULL,
  col_format = NULL,
  path_bib = "references.bib"
)
}
\arguments{
\item{dat}{A data frame containing the column to transform.}

\item{col_format}{A string specifying the name of the column to transform.}

\item{path_bib}{A file path to the bibliography file (\code{.bib}) used for citation keys.
Defaults to \code{"references.bib"} in the current working directory.}
}
\value{
A data frame with the specified column transformed.
}
\description{
This function applies the \link{ngr_cite_keys_to_inline} transformation to a specified column
in a data frame. It validates the inputs, checks that the specified column exists in the
data frame, and replaces its contents with the transformed values.
}
\examples{
# Path to a BibTeX file included in the package
path_bib <- system.file("extdata", "references.bib", package = "ngr")

# Example data frame
dat <- data.frame(
  id = 1:3,
  bib_keys = c(
    "Outside block quotes @busch_etal2011LandscapeLevelModela and in [@woll_etal2017SalmonEcological]",
    "This is many [@busch_etal2011LandscapeLevelModela; @woll_etal2017SalmonEcological; @kirsch_etal2014Fishinventory]",
    "this is a failed key @key3")
)

# Process the data frame
result <- ngr_cite_keys_to_inline_table_col(dat, col_format = "bib_keys", path_bib = path_bib)
result

}
\seealso{
Other cite: 
\code{\link{ngr_cite_assemble}()},
\code{\link{ngr_cite_bib_keys_missing}()},
\code{\link{ngr_cite_format}()},
\code{\link{ngr_cite_format_single}()},
\code{\link{ngr_cite_keys_extract}()},
\code{\link{ngr_cite_keys_to_xref}()}
}
\concept{cite}
