% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xct_keys_to_xref.R
\name{xct_assemble}
\alias{xct_assemble}
\title{Assemble Multiple Citations with APA Formatting}
\usage{
xct_assemble(citations)
}
\arguments{
\item{citations}{A vector of citation strings.}
}
\value{
A single string with all citations separated by semicolons.
}
\description{
Assembles a list of citations into a single inline reference, using semicolons.
}
\seealso{
Other cite: 
\code{\link{xct_bib_keys_missing}()},
\code{\link{xct_format}()},
\code{\link{xct_format_single}()},
\code{\link{xct_keys_extract}()},
\code{\link{xct_keys_to_inline}()},
\code{\link{xct_keys_to_xref}()}
}
\concept{cite}
