% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xct_keys_to_xref.R
\name{xct_format}
\alias{xct_format}
\title{Format Multiple Citations}
\usage{
xct_format(raw_key, bib_obj, ...)
}
\arguments{
\item{raw_key}{A raw citation key string, possibly with brackets.}

\item{bib_obj}{The bibliography object.}

\item{...}{Additional arguments passed from the main function.}
}
\value{
Formatted citation as a string.
}
\description{
Formats one or more citation keys into a single inline reference.
}
\seealso{
Other cite: 
\code{\link{xct_assemble}()},
\code{\link{xct_bib_keys_missing}()},
\code{\link{xct_format_single}()},
\code{\link{xct_keys_extract}()},
\code{\link{xct_keys_to_inline}()},
\code{\link{xct_keys_to_xref}()}
}
\concept{cite}
