% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xct_keys_to_inline.R,
%   R/xct_keys_to_inline_table_col.R
\name{xct_keys_to_inline}
\alias{xct_keys_to_inline}
\alias{xct_keys_to_inline_table_col}
\title{Convert Citation Keys to Inline References}
\usage{
xct_keys_to_inline(text, path_bib, csl_file = NULL)

xct_keys_to_inline_table_col(
  dat = NULL,
  col_format = NULL,
  path_bib = "references.bib"
)
}
\arguments{
\item{text}{A character vector containing text with citation keys (e.g., \verb{@doe2020}).}

\item{path_bib}{A file path to the bibliography file (\code{.bib}) used for citation keys.
Defaults to \code{"references.bib"} in the current working directory.}

\item{csl_file}{A string specifying the path to the CSL file (e.g., \code{apa.csl}). Default is \code{NULL},
which uses Pandoc's default citation style.}

\item{dat}{A data frame containing the column to transform.}

\item{col_format}{A string specifying the name of the column to transform.}
}
\value{
A character vector with inline references replacing the citation keys.

A data frame with the specified column transformed.
}
\description{
This function processes a block of text containing citation keys (e.g., \verb{@doe2020})
and converts them to inline references using Pandoc. The function supports
custom bibliography and CSL (Citation Style Language) files.

This function applies the \link{xct_keys_to_inline} transformation to a specified column
in a data frame. It validates the inputs, checks that the specified column exists in the
data frame, and replaces citation keys with inline citations.
}
\examples{
# Path to a BibTeX file included in the package
path_bib <- system.file("extdata", "references.bib", package = "xciter")

# Example data frame
dat <- data.frame(
  id = 1:3,
  bib_keys = c(
    "Outside block quotes @busch_etal2011LandscapeLevelModela and in [@woll_etal2017SalmonEcological]",
    "This is many [@busch_etal2011LandscapeLevelModela; @woll_etal2017SalmonEcological; @kirsch_etal2014Fishinventory]",
    "this is a failed key @key3")
)

# Process the data frame
result <- xct_keys_to_inline_table_col(dat, col_format = "bib_keys", path_bib = path_bib)
result

}
\seealso{
Other cite: 
\code{\link{xct_bib_keys_missing}()},
\code{\link{xct_keys_extract}()}
}
\concept{cite}
