% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xct_keys_to_xref.R
\name{xct_keys_to_xref}
\alias{xct_keys_to_xref}
\title{Generate Cross-reference Data Frame with Original Keys in One Column and Inline APA-Style Citations in Another.}
\usage{
xct_keys_to_xref(
  citation_keys,
  path_bib,
  key_abort_on_missing = TRUE,
  key_check_response = "warn"
)
}
\arguments{
\item{citation_keys}{A character vector of citation keys.}

\item{path_bib}{Path to the .bib bibliography file.}

\item{key_abort_on_missing}{Logical. If TRUE, aborts on missing citation keys.}

\item{key_check_response}{Response level for missing fields in bib entries. Default is "warn". Passed to
\code{\link[RefManageR]{ReadBib}} in the RefManageR package}
}
\value{
A data frame with citation keys in one column and their formatted inline references in another.
}
\description{
We probably just want to use \code{\link[=xct_keys_to_inline]{xct_keys_to_inline()}} or \code{\link[=xct_keys_to_inline_table_col]{xct_keys_to_inline_table_col()}}but this will
This function requires the \code{RefManageR} package. If it is not installed, you can install it with
\code{install.packages("RefManageR")}.
}
\details{
This function takes citation keys passed as a list and generates inline APA-style references for each item of the list.
}
\examples{
\dontrun{
path_bib <- "path/to/references.bib"
citation_keys <- c("@smith2000", "[@smith2001; @doe2005]")
xct_keys_to_xref(citation_keys, path_bib)
}
}
\seealso{
\code{\link[=xct_keys_to_inline]{xct_keys_to_inline()}}, \code{\link[=xct_keys_to_inline_table_col]{xct_keys_to_inline_table_col()}}

Other cite: 
\code{\link{xct_assemble}()},
\code{\link{xct_bib_keys_missing}()},
\code{\link{xct_format}()},
\code{\link{xct_format_single}()},
\code{\link{xct_keys_extract}()},
\code{\link{xct_keys_to_inline}()}
}
\concept{cite}
