% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_cite_keys_to_xref.R
\name{ngr_cite_keys_to_xref}
\alias{ngr_cite_keys_to_xref}
\title{We probably just want to use \link{ngr_cite_keys_table_col}  or \link{ngr_cite_keys_to_inline} but this will
Generate Cross-reference Data Frame with Original Keys in One Column and Inline APA-Style Citations in Another.}
\usage{
ngr_cite_keys_to_xref(
  citation_keys,
  bib_file,
  key_abort_on_missing = TRUE,
  key_check_response = "warn"
)
}
\arguments{
\item{citation_keys}{A character vector of citation keys.}

\item{bib_file}{Path to the .bib bibliography file.}

\item{key_abort_on_missing}{Logical. If TRUE, aborts on missing citation keys.}

\item{key_check_response}{Response level for missing fields in bib entries. Default is "warn". Passed to
\code{\link[RefManageR]{ReadBib}} in the RefManageR package}
}
\value{
A data frame with citation keys in one column and their formatted inline references in another.
}
\description{
This function takes citation keys passed as a list and generates inline APA-style references for each item of the list.
}
\examples{
\dontrun{
bib_file <- "path/to/references.bib"
citation_keys <- c("@smith2000", "[@smith2001; @doe2005]")
ngr_cite_keys_to_xref(citation_keys, bib_file)
}
}
\seealso{
Other cite: 
\code{\link{ngr_cite_assemble}()},
\code{\link{ngr_cite_format}()},
\code{\link{ngr_cite_format_single}()},
\code{\link{ngr_cite_keys_table_col}()}
}
\concept{cite}
