% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xct_bib_clean.R
\name{xct_bib_keys_missing}
\alias{xct_bib_keys_missing}
\alias{xct_bib_keys_extract}
\alias{xct_keys_guess_match}
\title{Check for Missing Citation Keys in a BibTeX File}
\usage{
xct_bib_keys_missing(path_bib, citations)

xct_bib_keys_extract(path_bib)

xct_keys_guess_match(keys_missing, keys_bib)
}
\arguments{
\item{path_bib}{A file path to the bibliography file (\code{.bib}) used for citation keys.
Defaults to \code{"references.bib"} in the current working directory.}

\item{keys}{\link{character} A vector of citation keys to check for in the BibTeX file.}
}
\value{
A vector of citation keys missing from the BibTeX file. Returns an empty
vector if all citation keys are found.

A character vector of citation keys extracted from the BibTeX file.

A character vector of citation keys extracted from the BibTeX file.
}
\description{
This function identifies citation keys provided in a list that are missing from
a specified BibTeX file. It ensures the BibTeX file exists and processes the
file to match the citation keys.

This function extracts all citation keys from a given BibTeX file.
}
\examples{
\dontrun{
path_bib <- system.file("extdata", "references.bib", package = "xciter")
citations <- c("Smith2020", "Jones2019")
xct_bib_keys_missing(path_bib, citations)
}
}
\seealso{
Other cite: 
\code{\link{xct_assemble}()},
\code{\link{xct_format}()},
\code{\link{xct_format_single}()},
\code{\link{xct_keys_extract}()},
\code{\link{xct_keys_to_inline}()},
\code{\link{xct_keys_to_xref}()}
}
\concept{cite}
